move bops mock data to shared file
